add dying support, query are awesome
I added the Death Component for when your dead , and i modify the query
for the player to exclude the death. Bam all my code support it now and
the only things remaining is that the animation must only run one time
for the dead, need to support it in my shitty system.